Add basic support for std::list<>.
This includes pretty printers for std::list<> and std::list<>
iterators as well as an xmethod for operator* of std::list<>
iterators.  The pretty printer for std::list<> iterators is kind of
redundant with the operator* since 'p *it' is what one would normally
use anyway since it matches the C++ syntax, but libstdc++ provides a
pretty printer for iterators, so match it.  I had an operator->
xmethod for the iterators, but it did not fully work.  Also, in my
simple testing wasn't needed, so I left it out.